+++ src/Products/Five/viewlet/metaconfigure.py
@@ -20,17 +20,15 @@ import os
 from zope.app.publisher.browser import viewmeta
 from zope.component import zcml
 from zope.configuration.exceptions import ConfigurationError
-from zope.interface import classImplements
-from zope.interface import implements
 from zope.interface import Interface
 from zope.browser.interfaces import IBrowserView
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 from zope.viewlet import interfaces
 
+from AccessControl.security import protectClass
+from AccessControl.security import protectName
 from App.class_init import InitializeClass
-from App.special_dtml import DTMLFile
-from Products.Five.security import protectClass
-from Products.Five.security import protectName
+
 from Products.Five.viewlet import manager
 from Products.Five.viewlet import viewlet
 
@@ -156,9 +154,6 @@ def viewletDirective(
             cdict.update(kwargs)
             new_class = type(class_.__name__,
                              (class_, viewlet.SimpleAttributeViewlet), cdict)
-
-        if hasattr(class_, '__implements__'):
-            classImplements(new_class, IBrowserPublisher)
 
     else:
         # Create a new class for the viewlet template alone.